Moving UpdateDisplaceMeshThread to it's own file closes #2812
