+++ framework/src/base/DisplacedProblem.C
@@ -16,6 +16,7 @@
 #include "Problem.h"
 #include "SubProblem.h"
 #include "ExodusOutput.h"
+#include "UpdateDisplacedMeshThread.h"
 
 template<>
 InputParameters validParams<DisplacedProblem>()
@@ -25,86 +26,6 @@ InputParameters validParams<DisplacedProblem>()
   params.addPrivateParam<bool>("sequence", true);
   return params;
 }
-
-class UpdateDisplacedMeshThread
-{
-public:
-  UpdateDisplacedMeshThread(DisplacedProblem & problem) :
-      _problem(problem),
-      _ref_mesh(_problem.refMesh()),
-      _nl_soln(*_problem._nl_solution),
-      _aux_soln(*_problem._aux_solution)
-  {
-  }
-
-  void operator() (const SemiLocalNodeRange & range) const
-  {
-    ParallelUniqueId puid;
-
-    std::vector<std::string> & displacement_variables = _problem._displacements;
-    unsigned int num_displacements = displacement_variables.size();
-
-    std::vector<unsigned int> var_nums;
-    std::vector<unsigned int> var_nums_directions;
-
-    std::vector<unsigned int> aux_var_nums;
-    std::vector<unsigned int> aux_var_nums_directions;
-
-    for(unsigned int i=0; i<num_displacements; i++)
-    {
-      std::string displacement_name = displacement_variables[i];
-
-      if (_problem._displaced_nl.sys().has_variable(displacement_name))
-      {
-         var_nums.push_back(_problem._displaced_nl.sys().variable_number(displacement_name));
-         var_nums_directions.push_back(i);
-      }
-      else if (_problem._displaced_aux.sys().has_variable(displacement_name))
-      {
-         aux_var_nums.push_back(_problem._displaced_aux.sys().variable_number(displacement_name));
-         aux_var_nums_directions.push_back(i);
-      }
-      else
-        mooseError("Undefined variable '"<<displacement_name<<"' used for displacements!");
-    }
-
-    unsigned int num_var_nums = var_nums.size();
-    unsigned int num_aux_var_nums = aux_var_nums.size();
-
-    unsigned int nonlinear_system_number = _problem._displaced_nl.sys().number();
-    unsigned int aux_system_number = _problem._displaced_aux.sys().number();
-
-    SemiLocalNodeRange::const_iterator nd = range.begin();
-
-    for (nd = range.begin() ; nd != range.end(); ++nd)
-    {
-      Node & displaced_node = *(*nd);
-
-      Node & reference_node = _ref_mesh.node(displaced_node.id());
-
-      for(unsigned int i=0; i<num_var_nums; i++)
-      {
-        unsigned int direction = var_nums_directions[i];
-        if (reference_node.n_dofs(nonlinear_system_number, var_nums[i]) > 0)
-          displaced_node(direction) = reference_node(direction) + _nl_soln(reference_node.dof_number(nonlinear_system_number, var_nums[i], 0));
-      }
-
-      for(unsigned int i=0; i<num_aux_var_nums; i++)
-      {
-        unsigned int direction = aux_var_nums_directions[i];
-        if (reference_node.n_dofs(aux_system_number, aux_var_nums[i]) > 0)
-          displaced_node(direction) = reference_node(direction) + _aux_soln(reference_node.dof_number(aux_system_number, aux_var_nums[i], 0));
-      }
-    }
-  }
-
-protected:
-  DisplacedProblem & _problem;
-  MooseMesh & _ref_mesh;
-  const NumericVector<Number> & _nl_soln;
-  const NumericVector<Number> & _aux_soln;
-};
-
 
 DisplacedProblem::DisplacedProblem(FEProblem & mproblem, MooseMesh & displaced_mesh, InputParameters params) :
     SubProblem(mproblem.name() + "_disp", params),
